requirements.txt
# windows-curses for windows development

python-dotenv==1.0.1
requests==2.31.0
PyYAML==6.0.1
mysql-connector-python==8.3.0
dnspython==2.6.1
num2words==0.5.13
discord.py==2.3.1
discord.py-Console==0.1.2
pillow==10.2.0
pyarr==5.2.0
openai==1.14.3
aiomysql==0.2.0
aiohttp==3.9.3
redis==5.0.3
hiredis==2.3.2
tiktoken==0.6.0 # ChatGPT Token counting
# wavelink==3.3.0 # music player
mafic==2.10.0 # better music player
PyNaCl==1.5.0 # music player